refactor transit
