+++ routes/src/lib.rs
@@ -5,3 +5,4 @@ pub mod food;
 pub mod printing;
 pub mod spaces;
 pub mod tools;
+pub mod transit;
