+++ src/routes.rs
@@ -1,2 +1,1 @@
-pub mod transit;
 pub mod webcams;
