+++ types/src/lib.rs
@@ -2,3 +2,4 @@ pub mod contacts;
 pub mod dictionary;
 pub mod food;
 pub mod spaces;
+pub mod transit;
